refactoring code style
Correcting the color of in history list queries

refactor
