+++ components/History/ItemHistory.type.ts
@@ -1,0 +1,8 @@
+import { IHistory } from "../../pages";
+
+export interface ItemHistory {
+  entry: IHistory;
+  index: number;
+  handleEdit: (arg: IHistory) => void;
+  isThemeDark: boolean;
+}
