+++ components/History/index.ts
@@ -1,0 +1,1 @@
+export { HistorysList } from "./HistorysList";
